PCL_Python_Analysis/pyproject.toml
[tool.poetry]
name = "Neuvisys"
version = "0.1.0"
description = ""
authors = ["Thomas Barbier"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
numpy = "^1.21.2"
matplotlib = "^3.4.3"
pandas = "^1.3.3"
dv = "^1.0.10"
Pillow = "^8.3.2"
natsort = "^7.1.1"
scipy = "^1.7.1"
scikit-learn = "^1.0"
seaborn = "^0.11.2"
opencv-python = "^4.5.3"
pyrosbag = "^0.1.3"
PyYAML = "^5.4.1"
pycryptodomex = "^3.11.0"
gnupg = "^2.3.1"
rospkg = "^1.3.0"
fpdf = "^1.7.2"
bitarray = "^2.3.4"
h5py = "^3.4.0"
jupyterlab = "^3.2.8"
notebook = "^6.4.8"
pdf2image = "^1.16.0"
neo = "^0.10.0"
viziphant = "^0.1.0"

[tool.poetry.dev-dependencies]
Sphinx = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
